Change evidence note compliance year to be int to match datetime.year

+++ src/EA.Weee.RequestHandlers.Tests.Unit/AatfEvidence/Requests/GetEvidenceNotesByOrganisationRequestTests.cs
@@ -60,7 +60,7 @@
         [Theory]
         [InlineData(0)]
         [InlineData(-1)]
-        public void GetEvidenceNotesByOrganisationRequest_GivenNComplianceYearIsNotGreaterThanZero_ArgumentOutOfRangeExceptionExpected(int complianceYear)
+        public void GetEvidenceNotesByOrganisationRequest_GivenNComplianceYearIsNotGreaterThanZero_ArgumentOutOfRangeExceptionExpected(short complianceYear)
         {
             //act
             var exception = Record.Exception(() =>
@@ -76,7 +76,7 @@
             //arrange
             var organisationId = fixture.Create<Guid>();
             var statusList = fixture.CreateMany<NoteStatus>().ToList();
-            var complianceYear = fixture.Create<int>();
+            var complianceYear = fixture.Create<short>();
 
             //act
             var result = new GetEvidenceNotesByOrganisationRequest(organisationId, statusList, complianceYear);
